修复数据库调用错误
