+++ src/plugins/group/welcome_switch.py
@@ -9,7 +9,7 @@ from nonebot.params import CommandArg
 
 require("nonebot_plugin_orm")
 
-from litebot_utils.models import commit_config, get_or_create_group_config
+from litebot_utils.models import get_or_create_group_config, get_session
 from litebot_utils.rule import is_group_admin
 from src.plugins.menu.models import MatcherData
 
@@ -35,19 +35,24 @@ async def _(
     # 获取当前群组的开关状态
     gid = event.group_id
     str_arg = arg.extract_plain_text().strip()
-    group_config, _ = await get_or_create_group_config(group_id=gid)
-    # 切换开关状态
-    if not str_arg:
-        await matcher.send(
-            f"成员变动提醒已 {'开启' if group_config.welcome else '关闭'} ！"
-        )
-    elif str_arg in ("on", "enable", "开启"):
-        group_config.welcome = True
-        await commit_config(group_config)
-        await matcher.send("成员变动提醒已开启！")
-    elif str_arg in ("off", "disable", "关闭"):
-        group_config.welcome = False
-        await commit_config(group_config)
-        await matcher.send("成员变动提醒已关闭！")
-    else:
-        await matcher.finish("请输入 on/off 来开启或关闭！")
+
+    async with get_session() as session:
+        group_config, _ = await get_or_create_group_config(group_id=gid)
+        # 切换开关状态
+        if not str_arg:
+            await matcher.send(
+                f"成员变动提醒已 {'开启' if group_config.welcome else '关闭'} ！"
+            )
+        elif str_arg in ("on", "enable", "开启"):
+            group_config.welcome = True
+            session.add(group_config)
+            await session.commit()
+            await matcher.send("成员变动提醒已开启！")
+        elif str_arg in ("off", "disable", "关闭"):
+            group_config.welcome = False
+
+            session.add(group_config)
+            await session.commit()
+            await matcher.send("成员变动提醒已关闭！")
+        else:
+            await matcher.finish("请输入 on/off 来开启或关闭！")
